Many improvements - Description
#DateTime
- now it returns an int, not a double

#Order
- the constructor now doesn't require a list of products
- implemented operator+= to add products to productList_
- implemented operator<< for outputs

#Product
- pricePerPiece_ is now double (no longer int)
- implemented operator<< for outputs

#Store
- the constructor no longer requires a productList_ and orderList_
- implemented += to add orders / products in a store
- implemented showSolvedOrdersOlderThan(x days)

+++ StoreManager/StoreManager/DateTime.cpp
@@ -94,5 +94,5 @@ int DateTime::operator-(const DateTime &other) {
 	std::time_t firstTime = std::mktime(&firstDate);
 	std::time_t secondTime = std::mktime(&secondDate);
 
-	return abs(floor(std::difftime(firstTime, secondTime) / (60 * 60 * 24)));
+	return int(abs(floor(std::difftime(firstTime, secondTime) / (60 * 60 * 24))));
 }
